Get tests for AffiliateInfo & RWS running again
